修正１

+++ app/controllers/items_controller.rb
@@ -37,7 +37,9 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    redirect_to root_path if item.destroy
+    if current_user.id == item.user_id
+      redirect_to root_path if item.destroy
+    end
   end
 
   private
